darwin: auto generate

+++ internal/testfixture/ccgo_darwin_amd64.go
@@ -50584,7 +50584,7 @@ func _postToParent(tls *libc.TLS, p uintptr, pScript uintptr) {
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript = pEvent + 1*32
 	libc.X__builtin___memcpy_chk(tls, (*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript, zMsg, uint64(*(*int32)(unsafe.Pointer(bp))+int32(1)), libc.X__builtin_object_size(tls, (*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript, 0))
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).Finterp = (*TSqlThread)(unsafe.Pointer(p)).Finterp
-	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, int32(_TCL_QUEUE_TAIL))
+	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, _TCL_QUEUE_TAIL)
 	libtcl8_6.XTcl_ThreadAlert(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent)
 }
 
@@ -50726,7 +50726,7 @@ func _sqlthread_parent(tls *libc.TLS, clientData TClientData, interp uintptr, ob
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript = pEvent + 1*32
 	libc.X__builtin___memcpy_chk(tls, (*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript, zMsg, uint64(*(*int32)(unsafe.Pointer(bp))+int32(1)), libc.X__builtin_object_size(tls, (*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript, 0))
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).Finterp = (*TSqlThread)(unsafe.Pointer(p)).Finterp
-	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, int32(_TCL_QUEUE_TAIL))
+	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, _TCL_QUEUE_TAIL)
 	libtcl8_6.XTcl_ThreadAlert(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent)
 	return m_TCL_OK
 }
